move set_classifier to contrib
